utils/requirements.txt
python-dotenv
openai
python-frontmatter
requests
